chore: states for uploads
