+++ warehouse/router/identities.go
@@ -448,7 +448,7 @@ func (r *Router) populateHistoricIdentities(ctx context.Context, warehouse model
 			return
 		}
 
-		_ = job.setUploadStatus(UploadStatusOpts{Status: getInProgressState(model.ExportedData)})
+		_ = job.setUploadStatus(UploadStatusOpts{Status: inProgressState(model.ExportedData)})
 		loadErrors, err := job.loadIdentityTables(true)
 		if err != nil {
 			r.logger.Errorf(`[WH]: Identity table upload errors: %v`, err)
