+++ warehouse/router/state.go
@@ -1,0 +1,105 @@
+package router
+
+import (
+	"fmt"
+
+	"github.com/rudderlabs/rudder-server/warehouse/internal/model"
+)
+
+var stateTransitions map[string]*state
+
+type state struct {
+	inProgress string
+	failed     string
+	completed  string
+	nextState  *state
+}
+
+func init() {
+	initializeStateMachine()
+}
+
+func initializeStateMachine() {
+	stateTransitions = make(map[string]*state)
+
+	waitingState := &state{
+		completed: model.Waiting,
+	}
+	stateTransitions[model.Waiting] = waitingState
+
+	generateUploadSchemaState := &state{
+		inProgress: "generating_upload_schema",
+		failed:     "generating_upload_schema_failed",
+		completed:  model.GeneratedUploadSchema,
+	}
+	stateTransitions[model.GeneratedUploadSchema] = generateUploadSchemaState
+
+	createTableUploadsState := &state{
+		inProgress: "creating_table_uploads",
+		failed:     "creating_table_uploads_failed",
+		completed:  model.CreatedTableUploads,
+	}
+	stateTransitions[model.CreatedTableUploads] = createTableUploadsState
+
+	generateLoadFilesState := &state{
+		inProgress: "generating_load_files",
+		failed:     "generating_load_files_failed",
+		completed:  model.GeneratedLoadFiles,
+	}
+	stateTransitions[model.GeneratedLoadFiles] = generateLoadFilesState
+
+	updateTableUploadCountsState := &state{
+		inProgress: "updating_table_uploads_counts",
+		failed:     "updating_table_uploads_counts_failed",
+		completed:  model.UpdatedTableUploadsCounts,
+	}
+	stateTransitions[model.UpdatedTableUploadsCounts] = updateTableUploadCountsState
+
+	createRemoteSchemaState := &state{
+		inProgress: "creating_remote_schema",
+		failed:     "creating_remote_schema_failed",
+		completed:  model.CreatedRemoteSchema,
+	}
+	stateTransitions[model.CreatedRemoteSchema] = createRemoteSchemaState
+
+	exportDataState := &state{
+		inProgress: "exporting_data",
+		failed:     "exporting_data_failed",
+		completed:  model.ExportedData,
+	}
+	stateTransitions[model.ExportedData] = exportDataState
+
+	abortState := &state{
+		completed: model.Aborted,
+	}
+	stateTransitions[model.Aborted] = abortState
+
+	waitingState.nextState = generateUploadSchemaState
+	generateUploadSchemaState.nextState = createTableUploadsState
+	createTableUploadsState.nextState = generateLoadFilesState
+	generateLoadFilesState.nextState = updateTableUploadCountsState
+	updateTableUploadCountsState.nextState = createRemoteSchemaState
+	createRemoteSchemaState.nextState = exportDataState
+	exportDataState.nextState = nil
+	abortState.nextState = nil
+}
+
+func inProgressState(state string) string {
+	uploadState, ok := stateTransitions[state]
+	if !ok {
+		panic(fmt.Errorf("invalid Upload state: %s", state))
+	}
+	return uploadState.inProgress
+}
+
+func nextState(dbStatus string) *state {
+	for _, uploadState := range stateTransitions {
+		if dbStatus == uploadState.inProgress || dbStatus == uploadState.failed {
+			return uploadState
+		}
+		if dbStatus == uploadState.completed {
+			return uploadState.nextState
+		}
+	}
+	return nil
+}
